move disk full tests to own module
